+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `AsyncError` helper to stich asynchronous exception stacktraces
+- Added `Dynalink` to any `DataObject`
+- Type definition for `process`
+- Loader updated to switch from `Nashorn`/`GraalVM` depending on the runtime
 - Fix Console `format` to not crash when handling unsupported types
 - Added a Changelog
 - Implemented `util/promisify` similar to node